Sample project: post list and JSON serialization
