+++ KakapoExample/KakapoExample.xcodeproj/xcshareddata/xcbaselines/DE9352951CAAF21300AD534C.xcbaseline/050470C2-82B5-4F27-AA37-056529EACDE9.plist
@@ -21,9 +21,9 @@
 				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
 				<dict>
 					<key>baselineAverage</key>
-					<real>0.26126</real>
+					<real>0.083181</real>
 					<key>baselineIntegrationDisplayName</key>
-					<string>14 apr 2016, 23:17:28</string>
+					<string>15 apr 2016, 18:47:54</string>
 				</dict>
 			</dict>
 			<key>testMultpleInsertions()</key>
@@ -34,6 +34,16 @@
 					<real>0.2452</real>
 					<key>baselineIntegrationDisplayName</key>
 					<string>14 apr 2016, 23:17:28</string>
+				</dict>
+			</dict>
+			<key>testMultpleInsertionsPerformance()</key>
+			<dict>
+				<key>com.apple.XCTPerformanceMetric_WallClockTime</key>
+				<dict>
+					<key>baselineAverage</key>
+					<real>0.080364</real>
+					<key>baselineIntegrationDisplayName</key>
+					<string>15 apr 2016, 18:47:54</string>
 				</dict>
 			</dict>
 		</dict>
