+++ KakapoExample/KakapoExample/AppDelegate.swift
@@ -7,15 +7,22 @@
 //
 
 import UIKit
+import Kakapo
 
 @UIApplicationMain
 class AppDelegate: UIResponder, UIApplicationDelegate {
 
     var window: UIWindow?
-
-
+    let db = KakapoDB()
+    
     func application(application: UIApplication, didFinishLaunchingWithOptions launchOptions: [NSObject: AnyObject]?) -> Bool {
-        // Override point for customization after application launch.
+        
+        setupRoutes(db)
+        
+        if let vc = window?.rootViewController as? ViewController {
+            vc.db = db
+        }
+        
         return true
     }
 
@@ -40,7 +47,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func applicationWillTerminate(application: UIApplication) {
         // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
     }
-
-
 }
