+++ KakapoExample/Podfile
@@ -12,4 +12,5 @@ end
 target 'KakapoExample' do
     pod 'Fakery'
     pod 'SnapKit'
+    pod 'SwiftyJSON'
 end
